wip: ProductSection 리팩토링중
