+++ src/refactoring/features/admin/components/product/ProductInfo.tsx
@@ -1,0 +1,30 @@
+import { formatDiscountRate } from "@/utils/formatDiscountRate";
+import { Product } from "@/types";
+
+interface Props {
+  product: Product;
+  handleEditProduct: (product: Product) => void;
+}
+
+export const ProductInfo = ({ product, handleEditProduct }: Props) => {
+  return (
+    <div>
+      {product.discounts.map((discount, index) => (
+        <div key={index} className="mb-2">
+          <span>
+            {discount.quantity}개 이상 구매 시{" "}
+            {formatDiscountRate(discount.rate)}
+            할인
+          </span>
+        </div>
+      ))}
+      <button
+        data-testid="modify-button"
+        onClick={() => handleEditProduct(product)}
+        className="bg-blue-500 text-white px-2 py-1 rounded hover:bg-blue-600 mt-2"
+      >
+        수정
+      </button>
+    </div>
+  );
+};
