+++ src/refactoring/features/cart/hooks/index.ts
@@ -1,6 +1,4 @@
 export * from './useCart';
-export * from './useCoupon';
-export * from './useProduct';
 export * from './useDiscount';
 export * from './useLocalStorage';
 export * from './useCartItem';
