+++ src/refactoring/features/shared/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './useCoupon';
+export * from './useProduct';
+
